requirements.txt
asgiref==3.10.0
Django==5.2.7
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
sqlparse==0.5.3
gunicorn==20.1.0
celery==5.3.1
redis==5.3.1
requests==2.31.0
